fix import of .pro files with CONFIG-=flat
With CONFIG-=flat in a .pro file "qmake -tp vc" generated the
solution folder "Generated Files/debug". On import the add-in tries
to created "Generated Files/Debug" because the subfolder "debug" is
not existent. But the folder creation failed because this is done
case insensitive.

QtProject.AddFileInSubfilter is now a little bit more fuzzy when
looking for existent solution subfolders.

Task-number: QTBUG-13830