Disable execution of VCCodeSign tool
A VS2005 or VS2008 MFC or other smart device project does not execute
teh CodeSign tool. If it does execute, it generates a -1 error on
command ''. If you examine Authenticode Config in project options, both
entries are No. Selecting the first No, restores the dropdown selections
for Authenticode. However, this clears AdditionalFiles. Setting
AdditionalFiles destroys the Authenticode setting. The only thing that
works is disabling the Authenticode tool in project tool build order.
This does this programmatically.

Merge-request: 1
Reviewed-by: Joerg Bornemann <joerg.bornemann@nokia.com>