fix batch builds for multiple platforms
Batch builds didn't work at all for multiple platforms.
The culprit was the attempt to select the solution's platform
whenever the Qt version of the project doesn't fit the build platform
(due to the call to ChangeQtVersion).
Since we save the Qt version per platform we don't have to do this at all.

Task-number: QTVSADDINBUG-51