Fix moc'ing problems.
Custom build step for moc'ing was many times run even not necessary.
Usually this happened if both 32b and 64b platform exist and there
were projects depending of each other in solution.

The root cause for this behaviour was that custom build command for
moc'ing changed when it shouldn't and MS build system triggered build
because of this.

Task-number: QTVSADDINBUG-92
Change-Id: I8b423ff4bd8f0686a7b2f14e68f7ba85b0b6bbd2
Reviewed-by: Markku Tapio Heikkilä <markku.heikkila@digia.com>