Fix for saving Qt5Version info to project globals.
When creating Qt Add-In project there is this 'Qt5Version' variable created
to project globals and set as persist. When converting project to QMake
generated one, this variable is not removed but persist is just set to false.

Now when trying to convert Qmake type project again to Qt Add-In this fails
because in changeset 997b0345 is a modification done which just checks the
existence of variable, not if it persists. And because of this persist is never
set and Add-In logic doesn't work correct anymore.

Adding persist check will fix the problem.

Old changeset 997b0345:

    diff --git a/Qt4VS2003/QtProjectLib/QtVersionManager.cs b/Qt4VS2003/QtProjectLib/QtVersionManager.cs
    index 6444b3f..58fd479 100644
    --- a/Qt4VS2003/QtProjectLib/QtVersionManager.cs
    +++ b/Qt4VS2003/QtProjectLib/QtVersionManager.cs
    @@ -244,9 +244,12 @@ namespace Nokia.QtProjectLib
        if (!IsVersionAvailable(version) && version != "$(DefaultQtVersion)")
            return false;
        string key = "QtVersion " + platform;
    -   project.Globals[key] = version;
    -   if (!project.Globals.get_VariablePersists(key))
    -       project.Globals.set_VariablePersists(key, true);
    +   if (!project.Globals.get_VariableExists(key) || project.Globals[key].ToString() != version)
    +   {
    +       project.Globals[key] = version;
    +       if (!project.Globals.get_VariablePersists(key))
    +           project.Globals.set_VariablePersists(key, true);
    +   }
        return true;
    }

Change-Id: I2a9aeb12fe05c2ed3ee37681eb672d990337ada8
Reviewed-by: Oliver Wolff <oliver.wolff@digia.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>