RefreshMocStep: Handle "standalone" header files containing Q_OBJECT properly
RefreshMocStep bailed out when it tried to obtain the cpp file's (that
does not exist) configuration in order to find includes, defines and pch
options. In order to be able to get these values (if applicable)
they are now fetched from the project configuration just like it is done
in AddMocStep.

Task-number: QTVSADDINBUG-399
Change-Id: I35ba5e29663e654e860fd3b3ec47d9be52eeda91
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>