Fixed addition of WEC2013 Qt versions
Since VS2010 it has not been possible to create a VCProjectEngine using
"new VCProjectEngineObject". The only way to obtain a project engine
is via currently open projects. As it is possible that the Qt options
dialog is shown while no project is open, we have to find another way of
obtaining the supported Windows CE platforms.

The needed information can be found in the SDK's Properties.xml. As the
installation paths of the Windows CE SDKs can be changed by the user,
they has to be obtained from the registry.

Task-numer: QTVSADDINBUG-422
Change-Id: I7173c3c2f2b09f6fd63e6b6ce17bbaa136148175
Reviewed-by: Karsten Heimrich <karsten.heimrich@theqtcompany.com>