Fix project import from .pro file
The mkspecs\\default directory does not exist anymore in never
versions of Qt and we could not extract and replace the Qt path
with the QTDIR environment variable. Relative paths are now replaced
under the assumption that the project is on the same drive as the
Qt version we do the import with. Fixes subsequent switches between
Qt versions e.g. from x86 to x64 etc...

Change-Id: I1a86ae3522535ae31cf47ee76fbc8f4aba8c6e4b
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>