Use of parallelism for Q_OBJECT parsing during .pro file import
First step to speed up the import of Qt projects from .pro files.
Execute parsing for Q_OBJECT, Q_GADET macros in parallel and move
it to a background task to not block the UI. This gains us close
to 50% on a capable machine in combination with the former patch.

Change-Id: Icb8acb1a633833dc490e94e3c9609894d30a7d4b
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>