Fix creating bad project if "flat" flag is removed
During .pro file import, RefreshRccSteps() now looks into subfilters
for rcc outputs to remove instead of just the top-level filter. In the
case where qmake generated a project structure with rcc outputs in the
"Generated Files/debug" and "Generated Files/release" filters (which is
the case with CONFIG -= flat), RefreshRccSteps() did not correctly
handle these outputs as it only looked for them in the topmost
"Generated Files" filter.

Task-number: QTVSADDINBUG-439
Change-Id: I2b71f0a511b2dd487f01f08fbc201e5f5006f72c
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>