Fix lupdate not working with Qt 5.9
When the command line to run lupdate was too long to specify individual
source files, a temporary .pro file was created instead. The processing
of the .pro file then triggered the same issue as QTVSADDINBUG-472
(qmake unable to find cl without VC vars). Removed the need for the
temporary .pro file by using the '@' option of lupdate (added a call to
SetVCVars() before running lupdate, just to be on the safe side).

Task-number: QTVSADDINBUG-491
Change-Id: I6bce4c1c72093eceee8d1951bbbe464c072ba1a3
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>