Replace custom build steps with MSBuild targets
Overhaul of Qt tool (moc, rcc, uic) usage in generated projects.
Running the Qt tools on project files no longer relies on custom build
steps but rather on Qt-specific MSBuild targets that are installed
together with the VS Tools. Older projects generated with previous
versions of the VS Tools can be converted to use the MSBuild targets
Conversion is not mandatory as custom build steps are still supported.

Change-Id: I4c01491b2c263deb2cdf2c5fcc12ecdf718eb11f
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>