Do not add generated files to project
Files generated by the Qt tools are no longer added to the project
structure. Source files generated by moc and rcc are included in the
build dynamically. This allows for an increase in performance of the VS
Tools (e.g. when adding header files to a project, 90% of the time is
spent processing generated files).
NOTE: Only works for new files/projects. This patch breaks import and
conversion of existing projects. Fixed in a following patch.

Task-number: QTVSADDINBUG-64
Change-Id: I8762adfc08962a54da256fe5ced6e1ec4504caaf
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>