Use XML transformation to convert to Qt/MSbuild
Convert projects based on custom build steps to Qt/MSBuild. The
converted projects do not include generated files for optimized
performance. The conversion uses only an XML transformation instead of
the Visual Studio SDK. This speeds up the import of .pro files,
especially for large projects as the performance of the VS SDK tends to
degrade with the number of calls to the COM interfaces.

Task-number: QTVSADDINBUG-442
Change-Id: If59a52f44eb2eb4dd99f386c5b695aa56025e900
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>