Read list of Qt modules from an XML file
Read a list of modules from an XML file. This list is used to setup the
module selection UI ("New Project" wizard and "Qt Project Settings"
dialog) and to configure preprocessor definitions, include paths and
link libraries required to use each module. This information was
previously hard-coded in several places.

Change-Id: Ie51006c20eb9aacd9ecc2d503b67fade566a15c7
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>