Check Qt target platform using DLL PE header
The function is64Bit() now looks into the PE header of Qt5Core.dll to
check if the target platform is 64 or 32 bit. Previously this relied on
the platform information of qmake.exe (obtained with GetBinaryType())
and so did not work for cross builds of Qt. This limitation was also
preventing the import of .pro files using a WinRT build of Qt.

Change-Id: Ib1af84c948d8bceb639601d3353482b3c92504fb
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>