Create UWP projects when using WinRT build of Qt
When creating a new Qt project (File > New > Project > Visual C++ > Qt)
if the default Qt version is a WinRT build, make the necessary
adjustments to the .vcxproj file to create a Universal Windows project.
Creating new ActiveQt Server, Console Application and Designer Custom
Widget projects is not allowed with a WinRT build of Qt.

Change-Id: If05f55a6df1a669460e3932ec9ebb5ab401016f6
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>