Fix getting unevaluated preprocessor definitions
When copying the preprocessor definitions from the compiler to moc, use
the IVCRulePropertyStorage interface to obtain fully evaluated data.
The previous method of iterating through property sheets and using the
VCCLCompilerTool interface is potentially slower and also may return
unevaluated information, as is the case when working with UWP projects.

Change-Id: I4b250016e9e2f7547030fc00a53dac72d4a0be73
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>