Fix issues with the use of the QtMsBuild environment variable
(1) When starting Visual Studio, set the environment variable for the
running process; this will enable VS Tools to work correctly in
side-by-side installations of Visual Studio.
(2) Remove the "ugly" (and now useless) registry access inside the
project file
(3) Don't define the useless environment variable "QtVsTools"
(4) If the Qt/MsBuild files cannot be found using the environment
variable, try a default location inside the project directory; this
will allow projects to be used "detached" from the VS Tools
(e.g. when sharing the project with someone that does not have the VS
Tools installed).

Change-Id: I7199d665c5a3801dc8804503ef5d887c2ae3344b
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>