Fix showing error messages when importing .pro files
When a project is added to the solution, call an overload of the method
SetDebuggingEnvironment() that allows a configuration to be specified.
During import of .pro files, the output window will display several
messages with the indication "Could not get the active platform name."
At this point, when the project is still being created, there is no
active configuration yet, so the overload that omits the configuration
should not be used.

Change-Id: If5dbcedc4ee12ca3ab485e5ff994a534ef11aa54
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>