Docs: Remove dependency to Qt's macros.qdocconf
The macros.qdocconf from Qt's global documentation config includes a
macro that uses the value of QT_VERSION, which is not defined for
Qt VS Tools manual.

The documentation build therefore fails with
'error: Environment variable 'QT_VERSION' undefined'.

To work around this, load the macros we use from a local qdocconf
file.

Change-Id: Iaa045bb3523e13a962d39e04dd41497e14f63bfe
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>