Set default Windows 10 SDK on VS 2017/2015
Sets the WindowsTargetPlatformVersion property to the default Windows 10
SDK, if any is installed. If this property is not defined, Visual Studio
will assume SDK version 8.1 and will issue an error in case it's not
installed. At least for VS 2017, installation of the Windows SDK 8.1 is
not included by default, so there is a good chance of seeing this error.
On VS 2013 the WindowsTargetPlatformVersion property is not supported so
this patch will only target VS 2017 and 2015.

Task-number: QTVSADDINBUG-479
Change-Id: Ifa437a965251a84ef598d73d2e05219be9a9b883
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>