Fix adding Qt class not configuring the moc step
Adding a Qt class (Project menu > Add > Add Qt Class...) did not
configure the header file correctly. The file would be labeled as custom
build and have an empty definition. This was due to an unconditional
call to EnsureCustomBuildToolAvailable (not applicable when using
Qt/MSBuild) together with an exception thrown trying to evaluate a
project-level property on a file-level configuration.

Change-Id: I69bfac3b95621ea68d8df2c242373b26b122720c
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>