Fix not using Qt installation if built with -qtlibinfix
Checks if a Qt installation was built with -qtlibinfix and, if so, uses
the infix to obtain the correct file names of DLL's and library files.
Previously, the infix was not taken into account, which meant that the
additional libraries information was incorrectly generated and, in
v2.2.0, Qt versions built with -qtlibinfix were not usable.

Task-number: QTVSADDINBUG-506
Change-Id: I9540bb8e43d8bf97ac303d212f90e8349b9532d7
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>