Fix removing quotes when converting custom build steps
Added support for escaped quotes when parsing the command line of custom
build steps. This was missing and caused problems, e.g. when a macro in
the form ABC="xyz" was defined. Also adjusted the expansion of the moc
macro definition property into the corresponding argument as it was not
compatible with quoted values in macro definitions.

Task-number: QTVSADDINBUG-514
Change-Id: Iafa487399def32828e7026d533672a0ce2496a65
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>