Fix usage of $(IntDir) in the definition of QtWorkFilePath
Changed the definition of the QtWorkFilePath property, taking into
account the possibility that $(IntDir) is an absolute path. The property
QtWorkFilePath points to the work file used by the Qt/MSBuild targets
(e.g. to collect the list of moc/rcc/uic tasks to run in parallel). The
previous definition generated build errors if $(IntDir) was defined as
an absolute path.

Task-number: QTVSADDINBUG-511
Change-Id: Ib8aa863ccc01498dd0842d3ab580330d35162ff6
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>