Handle command line strings longer than 8191 chars
Modified MSBuild targets to allow an argument string with more than 8191
when calling moc, rcc and uic. The built-in argument string generator is
designed to work with cmd.exe, which limits the maximum length of the
arguments string. Since we are creating the processes via .NET framework
instead of invoking cmd.exe, this limitation does not apply.

Change-Id: Idc02f7a4339df828c760318bbe45ecb1565c7ccb
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>