Use VS macros when converting to Qt/MSBuild
Modified the conversion of custom build steps to Qt/MSBuild in order to
use the macros %(Filename)%(Extension), $(Configuration) and $(Platform)
instead of specific values. This ensures that the values of properties
defined at project level are generic and can be applied to any file that
is added later.

Task-number: QTVSADDINBUG-512
Change-Id: I44f8d6f9be14566dc1708392e370f1455250e3b6
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>