Disable dynamic C++ source if already included
When converting from custom build to Qt/MSBuild, if a moc header does
not have a corresponding generated file, disable the dynamic C++ source
for that moc header. This is needed for the case when the moc output
is #include'd in the associated source file instead of being compiled on
its own.

Task-number: QTVSADDINBUG-530
Change-Id: Iae136ff6a14c867ed3250aa255b0681b77bcc29e
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>