Get target architecture from qconfig.pri (QT_ARCH)
Enhanced the parser of qconfig.pri to be able to access the value of
QT_ARCH, including the case of cross-compilation (where this variable
is defined inside a conditional block). The previous method of parsing
the Qt5Core.dll PE record does not work with a static build of Qt.

Task-number: QTVSADDINBUG-523
Change-Id: Ic12e35c520c8af90ae7a44f8b7dd191cbbd7c408
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>