Fix delay saving .cpp file
For projects that still use custom build steps, optimized the procedure
that checks if .cpp files have a moc step associated. This now relies
only on the presence of a .cbt file, which is much faster than the
previous method of going through every file in the project, times every
configuration, trying to find the .cpp as source for a moc command. This
method was causing noticeable delays when saving any .cpp file in larger
projects created with older versions of the Qt VS Tools.

Task-number: QTVSADDINBUG-509
Change-Id: I6c0543ddb15e444c8c91838c76c8f98fadff0f93
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>