Improve performance of conversion to Qt/MSBuild
Modified the conversion from custom build steps to Qt/MSBuild such that
larger projects can be converted faster. The optimization focused on the
code that removes generated items from the project: instead of looking
at every item in the project, only .cpp files will now be checked if
they are generated sources and should therefore be excluded from the
project. To further speed up this process, added an index on the path of
.cpp files.
Additionally, to provide UI feedback during the conversion procedure, a
progress dialog is now displayed when converting whole solutions.

Task-number: QTVSADDINBUG-515
Change-Id: Ib9326554bcf99dbff8b6dd0c65d1a44e18ab68b1
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>