Disable "fast up-to-date check" by default
Disable by default the fast up-to-date check that is performed by Visual
Studio to decide whether or not to call MSBuild. Setting the property
DisableFastUpToDateCheck ensures an "accurate" build, fully based on
MSBuild semantics:
"Whether Visual Studio should do its own faster up-to-date check before
Building, rather than invoke MSBuild to do a possibly more accurate one.
You would set this to [true] if you have a heavily customized build
process and builds in Visual Studio are not occurring when they should."
(https://goo.gl/4riNrR)
In recent versions of VS 2017, the "fast up-to-date check" fails in the
case of the Qt/MSBuild targets, leading to projects not being built when
a .uic or .qrc file is changed.

Task-number: QTVSADDINBUG-531
Change-Id: Ib847fbdad881d747afcbc963144100a60825ae91
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>