Don't call SetQtEnvironment during debug session
During the event handler for the debug start command, only call
SetQtEnvironment when the debug session is about to start. This event
is fired every time the debug start command (F5) is executed, which
includes the case of resuming execution after hitting a breakpoint. In
this case, calling SetQtEnvironment is not needed and may cause
noticeable delays when debugging large projects.

Task-number: QTVSADDINBUG-532
Change-Id: I6a4725631185c3166907429222cc5a3d619c4cc2
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>