Rollback changes to project XML on errors
Public methods of the MsBuildProject class will only have side-effects
in case of success. Previously this was not the case, which caused
problems, e.g. in the import of .pro files: if an error occurred during
conversion to Qt/MSBuild, the project file was saved with an incorrect
format and the project could not be loaded.

Task-number: QTVSADDINBUG-533
Change-Id: I9e7ca279b4c53d08c07ee54240f2dc246a661f17
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>