Fix problems with incremental build
Added a log file with the options used to invoke moc, rcc and uic on
each of the Qt source files. This information is used to determine which
files need to be re-processed in case one or more options were changed.
Previously, there was no mechanism to accurately verify this. The
modified date of the project file was checked against that of the output
files and, if the project was more recent, that would trigger a re-build
of all moc, rcc and uic steps, even if the modifications were not
related to the Qt tools.

Task-number: QTVSADDINBUG-535
Change-Id: I0a856b48efeda9e8af405d458f58eb28bb8caf4f
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>