Fix incorrect escaping of paths in command line strings
When generating command line strings to invoke Qt tools, in options
specifying file system paths, if the option ends with a backslash,
it is now escaped properly by appending a further '\' (cf. the
documentation on the CommandLineToArgvW() Windows API function
https://msdn.microsoft.com/en-us/library/windows/desktop/bb776391).
Previously, this unescaped backslash would cause problems when the
option was enclosed in quotes (the terminating \" would be interpreted
as an escaped literal quotation mark).

Task-number: QTVSADDINBUG-538
Change-Id: I481d7f83226fb1e7c3352acc6d9f43771c7e788a
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>