Add configurable QML syntax highlighting
Integrated the parser in Qt Declarative with the Visual Studio editor
extensibility classes for the purpose of implementing a QML syntax
highlighting that is configurable, and has better performance, accuracy
and usability than previous approaches. In particular, this fixes an
issue reported in VS 2015 and VS 2017, where it is not possible to
configure the formatting of the QML syntax highlighting.

Task-number: QTVSADDINBUG-501
Change-Id: I8ed61d7a4dc0cd9eeeebc19a36d02f49b21957ac
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>