Add Qt Declarative integration
This patch adds a new C++ project to the Qt VS Tools repository. The
output is a DLL intended to be called from C# code in the Qt VS Tools
extension. The purpose of this DLL is to provide access to services in
the Qt Declarative module. Initially, this will focus on integrating the
QML parser in the syntax highlighting feature provided by the extension.
Further integration with Qt Declarative may be added in future patches.

Task-number: QTVSADDINBUG-501
Change-Id: If04efaca090c20766687b80e2875ad7d91e177fd
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>