Write Tlog entries for moc, rcc and uic
Added moc, rcc and uic entries to the file tracker log (Tlog) to ensure
that the up-to-date check is correct, cf. this reply from Microsoft:
https://developercommunity.visualstudio.com/solutions/184689/view.html

Also, removed the DisableFastUpToDateCheck property that was set to
"true" by default. This property was added to avoid inaccuracies in the
fast up-to-date check. These inaccuracies are resolved by the presence
of the new Tlog entries.

Task-number: QTVSADDINBUG-544
Change-Id: I21e70231d12a4191cca9ea6ad50aa5fbdd12711c
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>