Optimize QML syntax highlighting for larger files
The performance of the QML syntax highlighting was noticeably poor when
handling larger files (> 3000 lines); it took some time for the
highlighting to be displayed after the file had been opened, and also,
when scrolling down, the text editor would freeze for a moment.

The optimization is mostly encapsulated in the newly created
QmlAsyncClassifier class and auxiliary types. The approach was twofold:

1. Classifiers that need to parse the same version of a QML file will
now use a shared QML parser instance; only the first access will trigger
the parsing and will store the result of classification in a shared list
of tags. Subsequent accesses by other classifier objects will just read
the shared data. Previously, each classifier had its own instance of the
QML parser. Because VS will create several classifier objects for the
same file, this meant that the QML parser would be needlessly invoked
several times for the exact same content. This is what caused the delay
in displaying the syntax highlighting.

2. Access to the classification results is now achieved through a binary
search on the list of tags. Previously, this list was accessed linearly,
which caused a delay when the classifiers were queried by VS (e.g. when
scrolling down in the text editor).

Change-Id: I83544a9b03f964ff11e9ce3cbbde54304bb7ca03
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>