Use full path of dependencies when writing tracker log file
Recording file access log entries into the tracker files (*.tlog) now
uses the full path of additional dependency files. Previously, these
dependencies where added to the log with a path relative to the project,
which meant that the VS file tracker was unable to find them and would
not propertly manage incremental build, i.e. if a Qt item add any
additional dependencies, then the build would never be considered
up-to-date, and MSBuild would always be invoked, even if no changes were
made to any of the project files.

Change-Id: I614d9cb33e2e017f9e0efde760d7ed5ad672848d
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>