Disable fast initialization by default
Will now only run the package initialization in a background thread if
there is an environment variable %QTVS_INIT% set with value "fast".
Otherwise, initialization runs in the main thread, as before. This is
due to reports that initialization is unstable in version 2.2.2 and may
cause the extension not to start-up properly or at all.

Task-number: QTVSADDINBUG-571
Change-Id: I9ec76ac2fa41c1c2d60375f8462bb057269dbd99
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>