Implement QML debug protocol
Implemented the protocol for communicating with the QML debug runtime.
Given that this communication is based on the exchange of messages
formatted in JSON, the classes in the QtVsTools.Json namespace are used
to map the JSON data to corresponding data contracts, i.e. .NET classes
representing the parsed contents of the JSON messages. A data contract
class was defined for each type of message exchanged in the QML debug
protocol. These messages are then sent and received through interop with
the native QML debug client. A public API to the debug protocol is
provided in the QmlDebugger class.

Task-number: QTVSADDINBUG-412
Change-Id: Ifb40b8d2c310b24619887605a53a3c8c5b617b91
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>