Implement Visual Studio QML debug engine
Implemented a Visual Studio debug engine for QML using the debugger
extensibility interfaces of Active Debugging 7 (AD7). The debug engine
uses the QmlDebugger class to interface with the QML debug runtime.
Debug sessions are launched concurrently to C++ debug sessions and
triggered when detecting the start of Qt applications in debug mode,
if they are the output of Qt projects including QML resource files.

Task-number: QTVSADDINBUG-412
Change-Id: If3363092741c0c87e5293113ea4c33970f5842dc
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>