Add support for QML cache generation
Extended the handling of QRC files in Qt/MSBuild to enable the
generation of QML cache files in build time. This is the equivalent to
setting CONFIG+=qtquickcompiler in a qmake project. A dedicated property
page "QML Cache" is now available for rcc source files, where the
options passed to qmlcachegen.exe can be configured.

Task-number: QTVSADDINBUG-428
Change-Id: I94db583a1a416fb1d9a17b2bd8a2bf9e83186717
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>