Use local socket for QML debug connection
Switched to using local socket (named pipe) instead of TCP when
connecting to the QML debug runtime. Since the debuggee process is (at
least for the moment) only started locally, there is no reason to prefer
the use of TCP. On the other hand, using a local socket is recommended
when connecting to a process that is started locally, due to issues with
connection timeout and retry when using TCP, which may prevent the debug
session to start.

Change-Id: I6dbbf274412ba03602e2449b62b5b2b83a090e50
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>