Fix code analysis error
This patch fixes the following code analysis error:

    MSBUILD : error : CA0001 : An error was encountered while parsing IL
    for method: 'QtVsTools.Qml.Debug.AD7.InfoHelper`1+Mapping`2
    +<GetEnumerator>d__10.MoveNext', instruction at offset '0xD' with
    opcode 'Ldfld'.

This error, though not reported by the C# compiler -- which generates
IL code that is apparently correct -- is nevertheless reported by the
code analysis step that runs at the end of the Release build. This seems
due to a stricter interpretation of the 'yield return' statement.
Specifically, whether it is allowed in a method returning IEnumerator.

On the other hand, calling the list's own GetEnumerator instead of
iterating that list seems like the better approach.

Change-Id: I100bf1421f53d4a25e1fb9f04336bc3f5f78407c
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>