Connect to QML runtime even if no QRC is detected
In the QML debug engine, removed a check that required the presence of
QRC files in the project in order to connect to the QML runtime. This
was resulting in the QML runtime blocking forever waiting for connection
to the debug client, in case a QRC file was not used (e.g. QML files
were accessed from the file system directly).

Task-number: QTVSADDINBUG-599
Change-Id: I0b83c6e63560c24ee00b65c24214b05748db202b
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>