Don't show rcc warnings when listing resources
When running rcc at the start of build to list the contents of resource
files, redirect stderr to null to suppress any warnings. As rcc will be
invoked again as part of the actual build, this prevents the same
warning messages from being shown twice (e.g. in case of an empty file).

Change-Id: I149ad189e740b070f9becbbed4d8b6dae3312740
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>