Fix issues in error handling of conversion to Qt/MSBuild
When converting from custom build steps to Qt/MSBuild, errors will now
be properly handled: if a custom build item cannot be converted in one
of the configurations, an error/warning message will be printed in the
Qt VS Tools output window; if project conversion is not possible, a
message is shown and the project/solution is reloaded without any
changes. Previously, if an error occurred, there was no feedback about
the cause of the error. Also, when converting a whole solution, the
progress window would not be closed, leading the user to conclude that
the conversion "froze".

Task-number: QTVSADDINBUG-604
Change-Id: I02b4d967b1f73fa24f74f56cf111433585ec6546
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>