Fix incorrect dependency to C++ compiler targets
Set the Qt/MSBuild targets to depend on the BuildCompile target.
Previously, a dependency existed with the ClCompile target. However,
this was incorrect, as this target does not include all requirements for
a successful C++ build (e.g. the MakeDirsForCl target is not run,
meaning that output directories will not be created).

Task-number: QTVSADDINBUG-609
Change-Id: If79a748d878d72c796c7f00608d06d9f6cca1ec0
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>