Add support for VS 2019 platform toolset
The latest version of VS 2019 Preview now includes a new platform
toolset (v142). Projects created with VS 2019 should reference this
toolset and not the one from VS 2017 (v141).

Also, in VS 2019 it's now possible to reference the latest installed
version of the Win10 SDK by setting the WindowsTargetPlatformVersion
project property to "10.0".

Change-Id: Ie8c0f13e27cfdebe8977741cfb098ab3b2cb1919
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>