Add auto-test framework
This patch introduces an auto-test framework that allows
on-the-fly compilation and execution of snippets of C# code that
implement the automated testing of modules in Visual Studio. These
snippets -- called 'macros' -- are sent from a test client program (e.g.
a QtTest-enabled application) through a named pipe, and received by a
specialized Visual Studio macro server extension that processes and
executes the macros, and then reports back to the test client. A small
client application is included in the patch, for demonstration purposes.

Change-Id: I0c11332cb384298547083e06b92626ee0d855779
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>