Add Regex wrapper
Adding a new module to the Qt VS Tools solution, a class library that
provides a wrapper for .NET Regex objects. This module allows defining
regular expressions using C# statements instead of strings. These
statements evaluate to instances of the RegExpr class, which can then
be translated to the syntax used by .NET's Regex.

The pattern statements of RegExpr also allow the definition of tokens.
These are mapped to capture groups in Regex and may contain production
rules to transform captured text into instances of externally defined
types (e.g. an abstract syntax tree).

This module is intended to replace the use of .NET Regex in the
Qt VS Tools projects, with a view to making code more readable
and maintainable.

For further details, see included README file.

Change-Id: I6509eba7cac0e20a2abeea4f3d2960ff0cb02ab8
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>