Fix linker rule property evaluation
Using the GenerateDebugInformation field of the VCLinkerTool class is
restricted to cases where the underlying project property is either set
to 'true' or 'false'. With any other value -- namely 'DebugFastLink' or
'DebugFull' -- attempting to get the (bool) value of the field
GenerateDebugInformation will generate an exception. This would cause
the .pro file export to exclude projects that use non-boolean values
for this linker property.

For further details:

https://social.msdn.microsoft.com/Forums/SECURITY/en-US/ce6aabc6-45d0-44e2-a219-28ecd728924f/vclinkertoolgeneratedebuginformation-property-does-not-support-debugfull-and-debugfastlink?forum=vsx

Task-number: QTVSADDINBUG-598
Change-Id: Ia7c5608f89237a0eb5ca3b67e539b1ec67d98996
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>