Refactor Qt/MSBuild
** Reorganized the file structure of the MSBuild props and targets
files for Qt. Files pertaining to each Qt tool will now reside in a
corresponding subfolder of the Qt/MSBuild directory. In each subfolder,
the lists of props and targets files to import, as well as any
dependency to other tools, are defined in the file 'qt_import.props'.

** Renamed rule (i.e. property page) names to force the order in which
they will be listed in the project properties dialog.

** Other minor refactoring adjustments

Change-Id: I44d7e3bb106f9103b0bfb7328a1393f75d1576d8
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>