Edit Qt settings in property pages (V3 format)
Qt settings are now configured in the project property pages. This
includes the possibility to have different versions of the Qt settings
per project configuration. Previously, the Qt settings were edited in a
custom dialog, allowing only a single version of the settings that
would apply to all project configurations.

This change breaks the current version of the support for Qt VS Tools
projects. A new version (V3) of the project format is introduced. This
new format allows Qt settings to be stored in the same way as other
project properties. Previous versions will still work and a later change
will introduce the possibility to convert from previous versions to V3.

Task-number: QTVSADDINBUG-575
Change-Id: I96ac9fff6d6db387756448245f0b1a155b3ba4d6
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>