Remove hard-coded rule name
The add-in needs to know the names of Qt/MSBuild rules (i.e. property
pages) in order to query property values. Previously, rule names were
hard-coded, which meant that changes to those names needed to be
replicated in at least two locations (the rule XML file and the add-in
code). In this change, the rule names used by the add-in code are now
read from the rule XML files.

Change-Id: I1a36b61c489df34858c7ef44a6ebebc31f3ad505
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>