Add revision number based on git describe
Added a revision part to the extension version number. The revision
number is based on the output of git describe, i.e. the number of
commits from the last tag of a designated version (MAJOR.MINOR.PATCH).

The new format for revised versions will now be:

    MAJOR.MINOR.PATCH.REVISION

Non-revised versions (last tag is at HEAD) will not include REVISION.

The motivation for this change is the fact that the version support
in the Visual Studio Marketplace does not allow reverting to older
version numbers. In case of a revert or hot fix, the new revision number
will now make it possible to push revised extensions to the marketplace
without the need to change the designated version.

Change-Id: Ic55c3bcdfad47d386e324ad36492159d77a59b2f
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>