Fix compatibility issues with VS 2015
MSBuild in VS 2015 does not process rule definitions in the same way as
in VS 2017 and 2019. In particular, the rule extension feature does not
allow overriding of property definitions. Because of this, rule
extension is no longer used and different versions of the definitions
are now included. The correct definition is selected according to the
version of the project format.

Another difference in the VS 2015 MSBuild is the fact that <Import>
does not handle correctly project path strings with line breaks,
issuing an error when they are present (in VS 2017 and 2019, line
breaks at the end of paths are ignored).

Change-Id: I919171115a12923c2fa087175d003802b18da1a9
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>