Fix error when importing .pro with static lib
When converting projects to the latest V3 format, will now detect the
usage of Qt modules by the presence of link library references and
QT_<module>_LIB macro definitions. Previously, only the presence of the
release link libraries were considered, leading to problems if these
were not referenced (i.e. in the case of a static lib project).

Task-number: QTVSADDINBUG-681
Change-Id: I0dfd2bc27056af340ce0cf306f31ce9849e442b4
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>