Remove designtime.props from incremental build check
The design-time Qt properties file will no longer be part of the
criteria to trigger an incremental build. This properties file is only
used to update Intellisense and should not be part of the actual build
process logic. Previously, incremental builds would be incorrectly
triggered because of this file.

Task-number: QTVSADDINBUG-694
Change-Id: Ic81079b8a1589e74392a6a68a9bee6306c03fb46
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>